Script to check DCHealth
This script will check the health status of DCs in a domain. Checks
include DFSR Health, DCDIAG and REPADMIN as well as configurable
services and files on the DCs and DNS records
The configuration is in the config.xml.
MailBody.html contains the html body for the mail that will be sent by
the script. MailBody .html can be customized to any design as long as
the placeholders remain intact